[sk] Major bug fixes
